Added Create Account E2E Test
